variable renames, function renames, cleanup
