+++ src/main/java/drzhark/mocreatures/entity/IMoCEntity.java
@@ -95,6 +95,4 @@ public interface IMoCEntity {
     boolean getIsFlying();
 
     String getClazzString();
-
-    boolean startRidingPlayer();
 }
