+++ src/main/java/drzhark/mocreatures/entity/tameable/MoCEntityTameableAmbient.java
@@ -514,4 +514,23 @@ public class MoCEntityTameableAmbient extends MoCEntityAmbient implements IMoCTa
     public void setGestationTime(int time) {
         gestationtime = time;
     }
+
+    @Override
+    public boolean startRidingPlayer(EntityPlayer player) {
+        if (MoCTools.getEntityRidingPlayer(player) != null) {
+            return false; // Something is already riding this player.
+        }
+        boolean ret = super.startRiding(player);
+        if (ret) {
+            NBTTagCompound tag = player.getEntityData();
+            tag.setUniqueId("MOCEntity_Riding_Player", this.getUniqueID());
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean canRidePlayer() {
+        return false;
+    }
 }
