+++ src/main/java/drzhark/mocreatures/entity/tameable/MoCEntityTameableAnimal.java
@@ -543,4 +543,18 @@ public class MoCEntityTameableAnimal extends MoCEntityAnimal implements IMoCTame
     public void setGestationTime(int time) {
         gestationtime = time;
     }
+
+    @Override
+    public boolean startRidingPlayer(EntityPlayer player) {
+        if (MoCTools.getEntityRidingPlayer(player) != null) {
+            return false; // Something is already riding this player.
+        }
+        boolean ret = super.startRiding(player);
+        if (ret) {
+            NBTTagCompound tag = player.getEntityData();
+            tag.setUniqueId("MOCEntity_Riding_Player", this.getUniqueID());
+            return true;
+        }
+        return false;
+    }
 }
