+++ src/main/java/drzhark/mocreatures/entity/tameable/MoCEntityTameableAquatic.java
@@ -538,6 +538,25 @@ public class MoCEntityTameableAquatic extends MoCEntityAquatic implements IMoCTa
     }
 
     @Override
+    public boolean startRidingPlayer(EntityPlayer player) {
+        if (MoCTools.getEntityRidingPlayer(player) != null) {
+            return false; // Something is already riding this player.
+        }
+        boolean ret = super.startRiding(player);
+        if (ret) {
+            NBTTagCompound tag = player.getEntityData();
+            tag.setUniqueId("MOCEntity_Riding_Player", this.getUniqueID());
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean canRidePlayer() {
+        return false;
+    }
+
+    @Override
     public void setGestationTime(int time) {
         gestationtime = time;
     }
